Consolidated ImageToSound and SoundToImage
