+++ src/__init__.py
@@ -1,3 +1,2 @@
-import src.ImageToSound
-import src.SoundToImage
+import src.Converters
 import src.Timer
